+++ cookiecutter.json
@@ -1,12 +1,11 @@
 {
-  "project_name": "Fantastic Wonder App",
-  "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '_').replace('-', '_') }}",
+  "project_name": "",
   "project_prefix": "{{ cookiecutter.project_name.lower().replace(' ', '') }}",
+  "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '-') }}",
   "django_secret_key_local": "",
   "django_secret_key_staging": "",
   "docker_registry_url": "docker.gitlab.liip.ch",
   "docker_registry_path": "amboss/{{ cookiecutter.project_name }}",
-  "project_short_description": "Testing out using cookiecutter to generate new forge apps",
   "_copy_without_render": [
     "*.html",
     "*.md",
