+++ {{ cookiecutter.project_slug }}/.env.staging
@@ -22,7 +22,7 @@ STATIC_URL=/static/
 MEDIA_ROOT=/tmp/static/media
 MEDIA_URL=/media/
 LOCALE_PATHS=locale/
-CORS_ORIGIN_WHITELIST=http://{{cookiecutter.project_slug}}.bedev.liip.ch
+CORS_ORIGIN_WHITELIST=https://{{cookiecutter.project_prefix}}-dev.bedev.liip.ch
 ACCESS_TOKEN_LIFETIME=15
 REFRESH_TOKEN_LIFETIME=120
 EMAIL_URL=smtp://smtp.example.net
@@ -39,5 +39,5 @@ PGADMIN_CONFIG_MASTER_PASSWORD_REQUIRED=False
 #####################
 # FRONTEND SETTINGS #
 #####################
-VUE_APP_API_URL=http://{{cookiecutter.project_slug}}.bedev.liip.ch/api/v1/
+VUE_APP_API_URL=http://api.{{cookiecutter.project_prefix}}-dev.bedev.liip.ch/api/v1/
 VUE_APP_ENABLE_AXE=true
