+++ {{ cookiecutter.project_slug }}/frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "{{cookiecutter.project_slug}}",
+  "name": "{{cookiecutter.project_prefix}}",
   "version": "0.1.0",
   "private": true,
   "scripts": {
